refactor(useValidTestTitle): address review comments and remove regex options
