--- crates/biome_js_analyze/tests/specs/nursery/useValidTestTitle/optionsPatterns.js
@@ -1,9 +1,0 @@
-/* should generate diagnostics */
-// Matches required pattern
-it('should work', () => {});
-
-// Fails mustMatch
-it('does not work', () => {});
-
-// Fails mustNotMatch
-it('should 123', () => {});
