--- crates/biome_js_analyze/tests/specs/nursery/useValidTestTitle/optionsPatterns.options.json
@@ -1,16 +1,0 @@
-{
-	"$schema": "../../../../../../packages/@biomejs/biome/configuration_schema.json",
-	"linter": {
-		"rules": {
-			"nursery": {
-				"useValidTestTitle": {
-					"level": "warn",
-					"options": {
-						"mustMatch": ["should .*"],
-						"mustNotMatch": ["should [0-9]+"]
-					}
-				}
-			}
-		}
-	}
-}
